git-svn-id: svn+ssh://svn.freebsd.org/base/head@14506 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/sys/vmmeter.h
@@ -30,8 +30,8 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)vmmeter.h	8.1 (Berkeley) 6/2/93
- * $Id: vmmeter.h,v 1.11 1995/07/29 10:47:05 bde Exp $
+ *	@(#)vmmeter.h	8.2 (Berkeley) 7/10/94
+ * $Id: vmmeter.h,v 1.13 1996/02/25 08:50:38 hsu Exp $
  */
 
 #ifndef _SYS_VMMETER_H_
@@ -44,50 +44,51 @@ struct vmmeter {
 	/*
 	 * General system activity.
 	 */
-	unsigned v_swtch;	/* context switches */
-	unsigned v_trap;	/* calls to trap */
-	unsigned v_syscall;	/* calls to syscall() */
-	unsigned v_intr;	/* device interrupts */
-	unsigned v_soft;	/* software interrupts */
+	u_int v_swtch;		/* context switches */
+	u_int v_trap;		/* calls to trap */
+	u_int v_syscall;	/* calls to syscall() */
+	u_int v_intr;		/* device interrupts */
+	u_int v_soft;		/* software interrupts */
 	/*
 	 * Virtual memory activity.
 	 */
-	unsigned v_vm_faults;	/* number of address memory faults */
-	unsigned v_cow_faults;	/* number of copy-on-writes */
-	unsigned v_zfod;	/* pages zero filled on demand */
-	unsigned v_swapin;	/* swap pager pageins */
-	unsigned v_swapout;	/* swap pager pageouts */
-	unsigned v_swappgsin;	/* swap pager pages paged in */
-	unsigned v_swappgsout;	/* swap pager pages paged out */
-	unsigned v_vnodein;	/* vnode pager pageins */
-	unsigned v_vnodeout;	/* vnode pager pageouts */
-	unsigned v_vnodepgsin;	/* vnode_pager pages paged in */
-	unsigned v_vnodepgsout;	/* vnode pager pages paged out */
-	unsigned v_intrans;	/* intransit blocking page faults */
-	unsigned v_reactivated;	/* number of pages reactivated from free list */
-	unsigned v_pdwakeups;	/* number of times daemon has awaken from sleep */
-	unsigned v_pdpages;	/* number of pages analyzed by daemon */
-	unsigned v_dfree;	/* pages freed by daemon */
-	unsigned v_pfree;	/* pages freed by exiting processes */
-	unsigned v_tfree;	/* total pages freed */
+	u_int v_vm_faults;	/* number of address memory faults */
+	u_int v_cow_faults;	/* number of copy-on-writes */
+	u_int v_zfod;		/* pages zero filled on demand */
+	u_int v_swapin;		/* swap pager pageins */
+	u_int v_swapout;	/* swap pager pageouts */
+	u_int v_swappgsin;	/* swap pager pages paged in */
+	u_int v_swappgsout;	/* swap pager pages paged out */
+	u_int v_vnodein;	/* vnode pager pageins */
+	u_int v_vnodeout;	/* vnode pager pageouts */
+	u_int v_vnodepgsin;	/* vnode_pager pages paged in */
+	u_int v_vnodepgsout;	/* vnode pager pages paged out */
+	u_int v_intrans;	/* intransit blocking page faults */
+	u_int v_reactivated;	/* number of pages reactivated from free list */
+	u_int v_pdwakeups;	/* number of times daemon has awaken from sleep */
+	u_int v_pdpages;	/* number of pages analyzed by daemon */
+
+	u_int v_dfree;		/* pages freed by daemon */
+	u_int v_pfree;		/* pages freed by exiting processes */
+	u_int v_tfree;		/* total pages freed */
 	/*
 	 * Distribution of page usages.
 	 */
-	unsigned v_page_size;	/* page size in bytes */
-	unsigned v_page_count;	/* total number of pages in system */
-	unsigned v_free_reserved; /* number of pages reserved for deadlock */
-	unsigned v_free_target;	/* number of pages desired free */
-	unsigned v_free_min;	/* minimum number of pages desired free */
-	unsigned v_free_count;	/* number of pages free */
-	unsigned v_wire_count;	/* number of pages wired down */
-	unsigned v_active_count;/* number of pages active */
-	unsigned v_inactive_target; /* number of pages desired inactive */
-	unsigned v_inactive_count;  /* number of pages inactive */
-	unsigned v_cache_count;		/* number of pages on buffer cache queue */
-	unsigned v_cache_min;		/* min number of pages desired on cache queue */
-	unsigned v_cache_max;		/* max number of pages in cached obj */
-	unsigned v_pageout_free_min;	/* min number pages reserved for kernel */
-	unsigned v_interrupt_free_min;	/* reserved number of pages for int code */
+	u_int v_page_size;	/* page size in bytes */
+	u_int v_page_count;	/* total number of pages in system */
+	u_int v_free_reserved;	/* number of pages reserved for deadlock */
+	u_int v_free_target;	/* number of pages desired free */
+	u_int v_free_min;	/* minimum number of pages desired free */
+	u_int v_free_count;	/* number of pages free */
+	u_int v_wire_count;	/* number of pages wired down */
+	u_int v_active_count;	/* number of pages active */
+	u_int v_inactive_target; /* number of pages desired inactive */
+	u_int v_inactive_count;	/* number of pages inactive */
+	u_int v_cache_count;	/* number of pages on buffer cache queue */
+	u_int v_cache_min;	/* min number of pages desired on cache queue */
+	u_int v_cache_max;	/* max number of pages in cached obj */
+	u_int v_pageout_free_min;   /* min number pages reserved for kernel */
+	u_int v_interrupt_free_min; /* reserved number of pages for int code */
 };
 #ifdef KERNEL
 extern struct vmmeter cnt;
@@ -96,20 +97,20 @@ extern struct vmmeter cnt;
 /* systemwide totals computed every five seconds */
 struct vmtotal
 {
-	short	t_rq;		/* length of the run queue */
-	short	t_dw;		/* jobs in ``disk wait'' (neg priority) */
-	short	t_pw;		/* jobs in page wait */
-	short	t_sl;		/* jobs sleeping in core */
-	short	t_sw;		/* swapped out runnable/short block jobs */
-	long	t_vm;		/* total virtual memory */
-	long	t_avm;		/* active virtual memory */
-	long	t_rm;		/* total real memory in use */
-	long	t_arm;		/* active real memory */
-	long	t_vmshr;	/* shared virtual memory */
-	long	t_avmshr;	/* active shared virtual memory */
-	long	t_rmshr;	/* shared real memory */
-	long	t_armshr;	/* active shared real memory */
-	long	t_free;		/* free memory pages */
+	int16_t	t_rq;		/* length of the run queue */
+	int16_t	t_dw;		/* jobs in ``disk wait'' (neg priority) */
+	int16_t	t_pw;		/* jobs in page wait */
+	int16_t	t_sl;		/* jobs sleeping in core */
+	int16_t	t_sw;		/* swapped out runnable/short block jobs */
+	int32_t	t_vm;		/* total virtual memory */
+	int32_t	t_avm;		/* active virtual memory */
+	int32_t	t_rm;		/* total real memory in use */
+	int32_t	t_arm;		/* active real memory */
+	int32_t	t_vmshr;	/* shared virtual memory */
+	int32_t	t_avmshr;	/* active shared virtual memory */
+	int32_t	t_rmshr;	/* shared real memory */
+	int32_t	t_armshr;	/* active shared real memory */
+	int32_t	t_free;		/* free memory pages */
 };
 
 /*
@@ -132,22 +133,22 @@ struct vmtotal
 #define	NRMON	64
 
 /* data and stack size distribution counters */
-unsigned int	dmon[NDMON+1];
-unsigned int	smon[NSMON+1];
+u_int	dmon[NDMON+1];
+u_int	smon[NSMON+1];
 
 /* page in time distribution counters */
-unsigned int	pmon[NPMON+2];
+u_int	pmon[NPMON+2];
 
 /* reclaim time distribution counters */
-unsigned int	rmon[NRMON+2];
+u_int	rmon[NRMON+2];
 
 int	pmonmin;
 int	pres;
 int	rmonmin;
 int	rres;
 
-unsigned rectime;		/* accumulator for reclaim times */
-unsigned pgintime;		/* accumulator for page in times */
+u_int rectime;		/* accumulator for reclaim times */
+u_int pgintime;		/* accumulator for page in times */
 #endif
 
 #endif
